+++ tests/integration/daemon/daemon_plugin_performance.cpp
@@ -1,3 +1,0 @@
-#include <spdlog/spdlog.h>
-#include <yams/compat/unistd.h>
-#include <yams/daemon/client/daemon_client.h>